format code

+++ python/oneflow/test/modules/test_clamp.py
@@ -63,7 +63,9 @@ def _test_clamp_scalar_max(test_case, shape, device, dtype):
 
 
 def _test_clamp_integral(test_case, shape, device, dtype):
-    input = flow.tensor(np.random.randint(3, 10, shape), device=flow.device(device)).to(dtype)
+    input = flow.tensor(np.random.randint(3, 10, shape), device=flow.device(device)).to(
+        dtype
+    )
     of_out = flow.clamp(input, 1, 5)
     np_out = np.clip(input.numpy(), 1, 5)
     test_case.assertTrue(np.allclose(of_out.numpy(), np_out, 1e-05, 1e-05))
